misc/HtmlTestSuite/results/test-0934.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>HTMLTextAreaElement Test: select event</title>
<link href="http://www.intel.com/" rel="author" title="Intel"/>
<meta content="interact" name="flags"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

</head><body><p>Select any numberic characters in the text area below</p>

<form id="testForm" name="testForm">
  <textarea id="testtextarea">0123456789</textarea>
</form>

<script>

var textarea = document.getElementById("testtextarea");

setup({explicit_done : true});
setup({explicit_timeout : true});

on_event(textarea, "select", function(evt) {
  test(function() {
    assert_greater_than(textarea.value.substring(textarea.selectionStart, textarea.selectionEnd).length, 0, "Check if select event captured when text selected");
  });
  done();
});

</script>

<div id="log"/>
</body></html>
